refactor: relocate gossip code to separate module
