+++ portalnet/src/lib.rs
@@ -4,6 +4,7 @@ pub mod config;
 pub mod discovery;
 pub mod events;
 pub mod find;
+pub mod gossip;
 pub mod metrics;
 pub mod overlay;
 mod overlay_service;
